refactor(ai): consolidate preset store, generate flow, and human-loop chat
Extract preset-store/io, unify generate hooks, merge tool UI, settle agent
interjections on finish, and drop unused AI elements and structured chat tools.
