+++ src/components/upload-preset-dialog.tsx
@@ -5,13 +5,13 @@ import { Dialog, DialogContent, DialogDescription, DialogFooter, DialogHeader, D
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { toast } from "sonner";
+import type { PresetModel } from "@/lib/database";
+import { getPresetDisplayName } from "@/lib/preset-store";
 import {
-    getPresetDisplayName,
     getPresetDefaultFileName,
     getPresetUploadToken,
-    PresetModel,
     uploadPreset,
-} from "@/hooks/use-preset";
+} from "@/lib/preset-io";
 import { useMemo, useState } from "react";
 
 interface UploadPresetDialogProps {
