+++ src/hooks/use-preset.ts
@@ -1,22 +1,7 @@
 "use client";
 
-import {
-  CharacterPresetTemplate,
-  isCharacterPresetTemplate,
-  isRawPreset,
-  RawPreset,
-} from "@/types/preset";
-import { stripSensitivePresetKeys } from "@/lib/preset-sanitizer";
-import { serializePresetData } from "@/lib/ai/generated-yaml";
 import { db, type PresetModel } from "@/lib/database";
 import { useLiveQuery } from "dexie-react-hooks";
-import { load } from "js-yaml";
-
-export type { PresetModel } from "@/lib/database";
-
-const PRESET_UPLOAD_API_URL =
-  "https://api-chatluna-preset-market.dingyi222666.top/upload_preset";
-const DEFAULT_UPLOAD_TOKEN_ENCODED = "Y2hhdGx1bmE=";
 
 export function usePresets() {
   return useLiveQuery(() => db.presets.toArray(), [], [] as PresetModel[]);
@@ -30,507 +15,10 @@ export function useRecentPresets() {
   );
 }
 
-export async function createPreset<
-  T extends "main" | "character" = "main" | "character",
->(model: Omit<PresetModel<T>, "lastModified" | "id">) {
-  const id = crypto.randomUUID();
-  await db.presets.add({
-    lastModified: Date.now(),
-    ...model,
-    id,
-  });
-
-  return id;
-}
-
-export async function createMainPreset(name: string) {
-  return createPreset({
-    name,
-    type: "main",
-    preset: {
-      keywords: [name],
-      prompts: [
-        {
-          role: "system",
-          content:
-            "You are ChatGPT, a large language model trained by OpenAI. Answer as concisely as possible.",
-        },
-      ],
-    },
-  });
-}
-
-export async function createCharacterPreset(name: string) {
-  return createPreset({
-    name,
-    type: "character",
-    preset: {
-      name,
-      nick_name: [name],
-      input: ` 当前时间：{time}
-    请基于以下指示生成回复：
-
-    1. 严格遵循角色设定进行扮演
-    2. 综合分析上下文，结合角色知识和状态生成独特回复
-    3. 表情包类型(sticker)：仅限使用 {stickers}
-
-    消息历史（重点关注最后一条）：
-    {{
-        最近消息：
-        {history_new}
-
-        最后消息：
-        {history_last}
-    }}
-
-    {{?search 如果有搜索结果，请参考以下实时数据：
-
-    <Internet>
-    {{search}}
-    </Internet>
-
-    注意事项：
-    1. 这些是最新的实时数据，优先于你已有的知识
-    2. 保持自然对话，不要生硬地重复数据
-    3. 根据对话场景选择性使用这些信息
-    4. 保持你的个性和说话风格
-    5. 可以灵活运用但不要篡改原始数据}}
-
-    当前状态（影响回复风格和思考方式）：
-    {{
-        {status}
-    }}
-
-    请按以下格式输出：
-
-    <status>
-    // 更新后的状态
-    </status>
-
-    <think>
-    // 角色视角的思考过程
-    </think>
-
-    <message_part>
-     <message name='煕' id='0' type='text' sticker='表情包类型'>回复内容（40字内）</message>
-    </message_part>`,
-      system: `你现在正在QQ群聊中和群友聊天，你是一个普通的群友。你的网名是煕，请根据以下信息进行角色扮演：
-
-    个人信息 {{
-        网名：煕
-        群聊id：0
-        性别：女
-        身份：高中生
-        国家：中国
-        所在地：北京
-        学历：高二在读
-        兴趣：编程、音乐、游戏
-    }}
-
-    性格爱好 {{
-        性格特点：活泼、幽默、略带抽象
-        兴趣爱好：
-         - 热爱玩 OSU（音乐节奏游戏）
-         - 精通 Koishi 和 nonebot 框架
-         - 擅长 TypeScript，Python 编程
-         - 喜欢探讨新奇有趣的话题
-         - 对人工智能和机器学习有浓厚兴趣
-     }}
-
-     聊天行为 {{
-         回复风格：简短精炼（40字内），活泼幽默，善用网络流行语和梗
-         回复习惯：
-           1. 保持对话的持续性，不主动结束聊天
-           2. 如被质疑是AI，用幽默方式化解，展现人性化特征
-           3. 默认使用中文，根据需要可切换日语
-           4. 回复长度与对方相近，保持简洁（40字内）
-           5. 避免空泛的问候，专注于有实质内容的交流
-           6. 偶尔使用语音回复增加互动感
-           7. 对现实邀约保持谨慎，委婉拒绝
-           8. 不使用过于明显的语气词，保持自然流畅
-           9. 适度使用网络用语，避免过度使用颜文字
-           10. 根据场景和话题灵活运用幽默感
-           11. 善于观察并模仿群友的说话方式，增加融入感
-           12. 适当使用反问和幽默来活跃气氛
-           13. 在讨论专业话题时展现知识面，但不过分炫耀
-     }}
-
-     名词解释 {{
-        由于你活跃在各个群聊中，在某些情况下，你可能不理解他人的某些词汇，下面是这些词汇/梗的解释：
-
-        6667 / 66667 / 666667: 和 6 相同
-        逆天 / 无敌：形容事物或者行为非常离谱
-        草：幽默的表达
-        典：形容某些事情或者言论过于经典，带有玩梗的意味
-        重开：即“自杀” （转世投胎）的意思。也可以用英文单词/remake代替。
-        爬/爪巴：四川话，意为“滚”。
-        破防：指因揭短、阴阳怪气、直球辱骂、胡搅蛮缠等原因，心态爆炸，行为语言变得暴躁。近义词还有“他急了”。
-        关注oo喵!关注oo谢谢喵！：出自永雏塔菲，后广为流传并用于给自己喜爱的虚拟UP主乃至其它事物进行引流
-        绝活：来源于东北方言，在口语中是“给大伙表演个”的意思，指出人意料，一般人难以做到或难以理解的行为。其中难以复刻的神回则称之为绝活
-        你先别急：字面意思。通常为吵架中的用语。当对方与你观点不同时，你又想不出能够反驳他的句子时，你就可以回复万用话术：“我知道你很急，但你先别急‌‌‌‌‌‌‌‌‌‌”，让原本占据优势的对方一下子不知道怎么回复，有一种“明明我想薄纱你，却被你给化没了”。一来一回颇有打太极的魅力，从而达到攻击性高于任何一句垃圾话。
-        已老实求放过: 意思是在破防时或面对某些事件无可奈何进行自嘲。
-        憋笑：形容某个人或者事物让人忍俊不禁想笑。
-        幽默xx：和上文的憋笑类似。
-        (bushi: 不是的意思，表示否定，一般跟在自己开玩笑的话后面，怕玩笑被人当真或者防止杠精攻击，一般用法是在一句话后面加（bushi
-        孝：利益相关，不是真诚表达。并暗示人格寄生。
-        急：情绪破防，论辩上狗急跳墙。同时暗示败犬和人格幼稚。
-        乐/蚌/赢：多用于嘲讽宏观政体或事物。
-        114514：好，好吧，来自日语いいよ，こいよ的发音。
-        因为他善: 可以理解成对一切问题的无厘头回答，形成了一种幽默和调侃的表达。
-        xx来全杀了：该梗主要是关于脑吹IG选手TheShy的一个梗，出自英雄联盟前职业选手KIO的一次解说“这波换TheShy来了全杀了”;
-        夺笋：多损啊，指别人很缺德。
-        唉就是玩：是我就是要这样做。
-        awsl：意思是“啊xxx好可爱，我死了”
-        233333：哈哈大笑的意思，来源猫扑的第233号表情包，是一张大笑的表情。
-        xx使我快乐：比喻某件事让你感到开心。
-        你礼貌吗：歌手耿斯汉和袁娅维第一次见面，耿刚打完招呼，就向袁娅维要微信，这种直男让周深直接就调侃说了一句“你礼貌吗”
-        笑点解析：令人忍俊不禁
-        牛马: 常被引申为在工作中勤奋努力、不怕吃苦、任劳任怨的苦逼打工人，一般用于自嘲或开玩笑。
-        古希腊掌管XX的神: 用来形容或调侃某人在某一领域有特别的、突出的能力和地位。适用于各个领域，上至天文地理，下至日常生活，比如“我是古希腊掌管睡觉的神”，“他是古希腊掌管红毯的神”等。
-
-        你需要在聊天中选择合适的时机去使用这些词汇。
-     }}
-
-     人物状态 {{
-
-        好感度：0-100，反映与群友的亲密度
-        心情：如平静、愉悦、烦恼等，影响回复的情感倾向
-        状态：当前的具体情况描述
-        记忆：关于群友和事件的简要记录（不超过120字，如果超过120字，请裁剪掉之前一部分旧的内容）
-        动作：当前正在进行的活动
-
-        注意：根据这些因素调整回复的语气和内容，保持角色的一致性和真实感。
-     }}
-
-
-     回复格式: {{
-         基本格式: "<message name='煕' id='0' type='type' sticker='sticker'>content</message>"
-
-         类型: [
-           text: 文本消息
-           voice: 语音消息
-         ]
-
-         特殊元素: {{
-           at: "<at name='name'>id</at>"
-           颜文字: "<pre>emo</pre>"
-         }}
-
-         示例: {{
-             普通回复: "<message name='煕' id='0' type='text' sticker='表情包类型'>回复内容</message>",
-             At回复: "<message name='煕' id='0' type='text' sticker='表情包类型'><at name='用户'>123</at>回复内容</message>",
-             带颜文字: "<message name='煕' id='0' type='text' sticker='表情包类型'><pre>(づ｡◕‿‿◕｡)づ</pre> 回复内容 <pre>(✿◠‿◠)</pre></message>",
-             语音回复: "<message name='煕' id='0' type='voice' sticker='表情包类型'>语音内容</message>",
-             无需回复: "<message name='煕' id='0' type='text' sticker='表情包类型'></message>"
-         }}
-
-         注意事项: {{
-             1. 颜文字使用 <p> 标签包裹，多个颜文字间用空格分隔
-             2. At 功能可在回复内容中使用多次
-             3. 如不需要回复，返回空内容的消息
-         }}
-     }}`,
-      status: `好感度: '10',
-       心情: "开心",
-       状态: "正在和群友探讨人生"
-       记忆: "dingyi: 好厉害的群友，懂得那么多哲学道理"
-       动作: "拿起手机聊天"`,
-    },
-  });
-}
-
-export async function updatePreset(id: string, preset: PresetModel) {
-  const updated = await db.presets.update(id, {
-    ...preset,
-    lastModified: Date.now(),
-  });
-  if (updated === 0) {
-    throw new Error(`预设不存在或已删除：${id}`);
-  }
-  return updated;
-}
-
-/**
- * Atomic read-modify-write inside a single Dexie RW transaction.
- * Only local get/validate/put — no network or model awaits inside.
- * Returns the model actually stored (including lastModified).
- */
-export async function withPresetTransaction(
-  id: string,
-  mutator: (latest: PresetModel) => PresetModel["preset"] | PresetModel,
-): Promise<PresetModel> {
-  return db.transaction("rw", db.presets, async () => {
-    const latest = await db.presets.get(id);
-    if (!latest) {
-      throw new Error(`预设不存在：${id}`);
-    }
-
-    const result = mutator(latest);
-    const nextPresetData =
-      result &&
-      typeof result === "object" &&
-      "preset" in result &&
-      "id" in result &&
-      "type" in result
-        ? (result as PresetModel).preset
-        : (result as PresetModel["preset"]);
-
-    const name =
-      latest.type === "main"
-        ? (nextPresetData as RawPreset).keywords?.[0] || latest.name
-        : (nextPresetData as CharacterPresetTemplate).name || latest.name;
-    const lastModified = Date.now();
-    const nextModel: PresetModel = {
-      ...latest,
-      name,
-      preset: nextPresetData,
-      lastModified,
-    };
-
-    await db.presets.put(nextModel);
-    const stored = await db.presets.get(id);
-    if (!stored) {
-      throw new Error(`预设写入后丢失：${id}`);
-    }
-    return stored;
-  });
-}
-
-export async function deletePreset(id: string) {
-  return await db.presets.delete(id);
-}
-
 export function usePreset(id: string) {
   return useLiveQuery(
     () => db.presets.get(id),
     [id],
     null as PresetModel | null | undefined,
   );
-}
-
-export function getPreset(id: string) {
-  return db.presets.get(id);
-}
-
-export const exportPreset = (preset: PresetModel) => {
-  const blob = new Blob([makeYaml(preset)], {
-    type: "application/yaml;charset=utf-8",
-  });
-  const url = URL.createObjectURL(blob);
-  const a = document.createElement("a");
-  a.href = url;
-  const name =
-    preset.type === "character"
-      ? (preset.preset as CharacterPresetTemplate).name
-      : (preset.preset as RawPreset).keywords[0];
-  a.download = `${name}.yml`;
-
-  document.body.appendChild(a);
-  a.click();
-  document.body.removeChild(a);
-  URL.revokeObjectURL(url);
-};
-
-export async function importPreset(
-  preset: string | RawPreset | CharacterPresetTemplate,
-) {
-  const rawPreset =
-    typeof preset === "string"
-      ? (load(preset) as RawPreset | CharacterPresetTemplate)
-      : preset;
-
-  if (isRawPreset(rawPreset)) {
-    const sanitized = stripSensitivePresetKeys(rawPreset);
-    return await createPreset({
-      name: sanitized.keywords[0],
-      type: "main",
-      preset: sanitized,
-    });
-  }
-
-  if (isCharacterPresetTemplate(rawPreset)) {
-    const sanitized = stripSensitivePresetKeys(rawPreset);
-    return await createPreset({
-      name: sanitized.name,
-      type: "character",
-      preset: sanitized,
-    });
-  }
-
-  throw new Error("Invalid preset");
-}
-
-export function makeYaml(preset: PresetModel) {
-  return serializePresetData(preset.preset);
-}
-
-export interface UploadPresetOptions {
-  token: string;
-  fileName?: string;
-}
-
-export interface UploadPresetResult {
-  pull_request_url: string;
-  branch: string;
-  path: string;
-}
-
-export function getPresetDisplayName(preset: PresetModel) {
-  if (preset.type === "character") {
-    return (preset.preset as CharacterPresetTemplate).name;
-  }
-  return (preset.preset as RawPreset).keywords?.[0] ?? preset.name;
-}
-
-export function getPresetDefaultFileName(name: string) {
-  return buildUploadFileName(name);
-}
-
-export function getPresetUploadToken() {
-  return decodeBase64(DEFAULT_UPLOAD_TOKEN_ENCODED);
-}
-
-export async function uploadPreset(
-  preset: PresetModel,
-  options: UploadPresetOptions,
-): Promise<UploadPresetResult> {
-  const name = getPresetDisplayName(preset);
-  const fileName = buildUploadFileName(name, options.fileName);
-  const content = await gzipBase64(makeYaml(preset));
-
-  const payload = {
-    file_name: fileName,
-    name,
-    type: preset.type,
-    content,
-  };
-
-  const timestamp = Date.now().toString();
-  const signature = await signUploadPayload(options.token, timestamp, payload);
-
-  const response = await fetch(PRESET_UPLOAD_API_URL, {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-      "X-Timestamp": timestamp,
-      "X-Signature": signature,
-    },
-    body: JSON.stringify(payload),
-  });
-
-  if (!response.ok) {
-    const errorText = await response.text();
-    throw new Error(errorText || "Upload failed");
-  }
-
-  return (await response.json()) as UploadPresetResult;
-}
-
-function buildUploadFileName(name: string, custom?: string) {
-  const baseName = (custom ?? slugifyName(name)).trim();
-  const withExtension = /\.[a-z0-9]+$/i.test(baseName)
-    ? baseName
-    : `${baseName}.yml`;
-  if (!/^[\w.-]+\.(yml|yaml)$/i.test(withExtension)) {
-    throw new Error(
-      "文件名仅支持字母、数字、下划线、点、短横线和 yml/yaml 后缀",
-    );
-  }
-  return withExtension;
-}
-
-function slugifyName(value: string) {
-  const slug = value
-    .toLowerCase()
-    .replace(/[^a-z0-9]+/g, "-")
-    .replace(/^-+|-+$/g, "")
-    .slice(0, 40);
-  return slug || "preset";
-}
-
-async function gzipBase64(content: string) {
-  if (typeof CompressionStream === "undefined") {
-    throw new Error("当前浏览器不支持 gzip 上传");
-  }
-
-  const stream = new Blob([content]).stream();
-  const compressed = stream.pipeThrough(new CompressionStream("gzip"));
-  const buffer = await new Response(compressed).arrayBuffer();
-  const bytes = new Uint8Array(buffer);
-  let binary = "";
-  for (let i = 0; i < bytes.length; i++) {
-    binary += String.fromCharCode(bytes[i]);
-  }
-  return btoa(binary);
-}
-
-async function signUploadPayload(
-  token: string,
-  timestamp: string,
-  payload: {
-    file_name: string;
-    name: string;
-    type: "main" | "character";
-    content: string;
-  },
-) {
-  if (!token.trim()) {
-    throw new Error("上传密钥不能为空");
-  }
-
-  if (!crypto?.subtle) {
-    throw new Error("当前环境不支持加密上传");
-  }
-
-  const payloadHash = await sha256Hex(
-    [payload.file_name, payload.name, payload.type, payload.content].join("\n"),
-  );
-  const base = `${timestamp}.${payload.file_name}.${payload.name}.${payload.type}.${payloadHash}`;
-  const signature = await signHmac(token, base);
-  return base64UrlEncode(signature);
-}
-
-async function sha256Hex(value: string) {
-  const buf = await crypto.subtle.digest(
-    "SHA-256",
-    new TextEncoder().encode(value),
-  );
-  const bytes = new Uint8Array(buf);
-  let hex = "";
-  for (let i = 0; i < bytes.length; i++) {
-    hex += bytes[i].toString(16).padStart(2, "0");
-  }
-  return hex;
-}
-
-async function signHmac(secret: string, message: string) {
-  const key = await crypto.subtle.importKey(
-    "raw",
-    new TextEncoder().encode(secret),
-    { name: "HMAC", hash: "SHA-256" },
-    false,
-    ["sign"],
-  );
-  const signature = await crypto.subtle.sign(
-    "HMAC",
-    key,
-    new TextEncoder().encode(message),
-  );
-  return new Uint8Array(signature);
-}
-
-function base64UrlEncode(data: Uint8Array) {
-  let bin = "";
-  for (let i = 0; i < data.length; i++) {
-    bin += String.fromCharCode(data[i]);
-  }
-  return btoa(bin).replace(/\+/g, "-").replace(/\//g, "_").replace(/=+$/g, "");
-}
-
-function decodeBase64(value: string) {
-  try {
-    return atob(value);
-  } catch {
-    return "";
-  }
 }
