+++ src/types/preset.ts
@@ -63,60 +63,10 @@ export function isWorldLore(obj: RawWorldLore | WorldLoreConfig): obj is RawWorl
     );
 }
 
-export interface RoleBook {
-    keywords: (string | RegExp)[];
-    content: string;
-    scanDepth?: number;
-    recursiveScan?: boolean;
-    maxRecursionDepth?: number;
-    matchWholeWord?: boolean;
-    caseSensitive?: boolean;
-    enabled?: boolean;
-    constant?: boolean;
-    order?: number;
-    insertPosition?:
-        | "before_char_defs"
-        | "after_char_defs"
-        | "before_example_messages"
-        | "after_example_messages";
-}
-
-export type RoleBookConfig = Omit<PresetTemplate["loreBooks"], "items">;
-
 export interface BaseMessage {
     role: "user" | "system" | "assistant";
     type?: "personality" | "description" | "first_message" | "scenario";
     content: string;
-}
-
-export interface PresetTemplate {
-    version?: string;
-    triggerKeyword: string[];
-    rawText: string;
-    messages: BaseMessage[];
-    formatUserPromptString?: string;
-    path?: string;
-    loreBooks?: {
-        scanDepth?: number;
-        items: RoleBook[];
-        tokenLimit?: number;
-        recursiveScan?: boolean;
-        maxRecursionDepth?: number;
-        insertPosition?:
-            | "before_char_defs"
-            | "after_char_defs"
-            | "before_example_messages"
-            | "after_example_messages";
-    };
-    authorsNote?: AuthorsNote;
-    knowledge?: KnowledgeConfig;
-    config: {
-        longMemoryPrompt?: string;
-        loreBooksPrompt?: string;
-        longMemoryExtractPrompt?: string;
-        longMemoryNewQuestionPrompt?: string;
-        postHandler?: PostHandler;
-    };
 }
 
 export interface PostHandler {
@@ -136,19 +86,6 @@ export interface AuthorsNote {
     insertPosition?: "after_char_defs" | "in_chat";
     insertDepth?: number;
     insertFrequency?: number;
-}
-
-export function isRoleBook(obj: unknown): obj is RoleBook {
-    return (
-        typeof obj === "object" &&
-        obj !== null &&
-        "keywords" in obj &&
-        "content" in obj
-    );
-}
-
-export function isRoleBookConfig(obj: unknown): obj is RoleBookConfig {
-    return !isRoleBook(obj) && typeof obj === "object" && obj !== null;
 }
 
 export function isRawPreset(obj: unknown): obj is RawPreset {
